+++ pytest-webdriver/pytest_webdriver.py
@@ -65,6 +65,12 @@ def webdriver(request):
 
     if CONFIG.browser.lower() == 'phantomjs':
         driver = webdriver.PhantomJS(executable_path=CONFIG.phantomjs)
+    elif CONFIG.browser.lower() == 'chrome':
+        chrome_options = webdriver.ChromeOptions()
+        chrome_options.add_argument('--headless')
+        chrome_options.add_argument('--no-sandbox')
+        chrome_options.add_argument('--disable-dev-shm-usage')
+        driver = webdriver.Chrome(chrome_options=chrome_options)
     else:
         selenium_uri = CONFIG.uri
         if not selenium_uri: